+++ R/score_se.R
@@ -169,7 +169,9 @@ RMscoreSE <- function(data,
       x = "Logit interval score",
       y = "Ordinal sum score"
     ) +
-    ggplot2::theme_bw()
+    ggplot2::theme_bw() +
+    ggplot2::theme(axis.title.x = ggplot2::element_text(margin = ggplot2::margin(t = 12)),
+                   axis.title.y = ggplot2::element_text(margin = ggplot2::margin(r = 12)))
 }
 
 # ---------------------------------------------------------------------------