+++ NEWS.md
@@ -1,3 +1,12 @@
+# easyRasch2 0.5.6
+
+- Added **`RMresidualPCA()`** for evaluating patterns in the standardized residuals 
+  from the RM/PCM. Outputs either a table with eigenvalues and explained variance
+  or a figure with standardized loadings on the first residual contrast and item
+  locations.
+- Added **`RMpcaCutoff()`** to determine a simulation-based critical value for the
+  largest eigenvalue.
+ 
 # easyRasch2 0.5.5
 
 - Added **`RMbootRestscore()`** for use with large sample sizes. See <https://pgmj.github.io/rasch_itemfit/> for more details.